requirement.txt
pandas
addict
scikit-learn
opencv-python
pytorch_warmup
scikit-image
tqdm
termcolor
yapf